- Moved Project to sub folder - Added Application Insights
