+++ src/Team-Service-Bot.Api/Connected Services/Application Insights/ConnectedService.json
@@ -1,0 +1,7 @@
+﻿{
+  "ProviderId": "Microsoft.ApplicationInsights.ConnectedService.ConnectedServiceProvider",
+  "Version": "8.6.209.10",
+  "GettingStartedDocument": {
+    "Uri": "https://go.microsoft.com/fwlink/?LinkID=613413"
+  }
+}
