+++ src/Team-Service-Bot.Api/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Vsar.TSBot.WebApiApplication" Language="C#" %>
