+++ src/Team-Service-Bot.Api/Global.asax.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Http;
-using System.Web.Routing;
+﻿using System.Web.Http;
 
-namespace Team_Services_Bot
+namespace Vsar.TSBot
 {
     public class WebApiApplication : System.Web.HttpApplication
     {
