+++ src/Team-Service-Bot.Api/Properties/AssemblyInfo.cs
