+++ src/Team-Service-Bot.Api/Web.Release.config
