+++ src/Team-Service-Bot.Api/Web.config
@@ -22,6 +22,9 @@
     <customErrors mode="Off" /> 
     <compilation debug="true" targetFramework="4.6" />
     <httpRuntime targetFramework="4.6" />
+    <httpModules>
+      <add name="ApplicationInsightsWebTracking" type="Microsoft.ApplicationInsights.Web.ApplicationInsightsHttpModule, Microsoft.AI.Web" />
+    </httpModules>
   </system.web>
   <system.webServer>
     <defaultDocument>
@@ -36,7 +39,13 @@
       <remove name="OPTIONSVerbHandler" />
       <remove name="TRACEVerbHandler" />
       <add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0" />
-    </handlers></system.webServer>
+    </handlers>
+  <validation validateIntegratedModeConfiguration="false" />
+  <modules>
+  <remove name="ApplicationInsightsWebTracking" />
+  <add name="ApplicationInsightsWebTracking" type="Microsoft.ApplicationInsights.Web.ApplicationInsightsHttpModule, Microsoft.AI.Web" preCondition="managedHandler" />
+  </modules>
+  </system.webServer>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
