+++ src/Team-Service-Bot.Api/default.htm
@@ -5,7 +5,7 @@
     <meta charset="utf-8" />
 </head>
 <body style="font-family:'Segoe UI'">
-    <h1>Team_Services_Bot</h1>
+    <h1>Team Services Bot</h1>
     <p>Describe your bot here and your terms of use etc.</p>
     <p>Visit <a href="https://www.botframework.com/">Bot Framework</a> to register your bot. When you register it, remember to set your bot's endpoint to <pre>https://<i>your_bots_hostname</i>/api/messages</pre></p>
 </body>
