+++ src/Team-Service-Bot.Api/packages.config
@@ -2,6 +2,13 @@
 <packages>
   <package id="Autofac" version="3.5.2" targetFramework="net46" />
   <package id="Chronic.Signed" version="0.3.2" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights" version="2.2.0" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.0.6" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.2.0" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.2.0" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.Web" version="2.2.0" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.2.0" targetFramework="net46" />
+  <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.2.0" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net46" />
