--- src/Team-Services-Bot/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Team_Services_Bot.WebApiApplication" Language="C#" %>
